deepcheck_backend/requirements.txt
aiohttp==3.7.4.post0
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==20.1.0
async-timeout==3.0.1
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
colored==1.4.4
colorful==0.5.5
cryptography==3.4.8
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
environs==9.3.5
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.12.0
httpcore==0.13.7
httptools==0.6.0
httpx==0.19.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
merge-args==0.1.5
multidict==6.0.4
packaging==23.1
passlib==1.7.4
prettyprinter==0.18.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-core==2.4.0
Pygments==2.15.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.23.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.16.2
